Languages checkpoint 4

+++ core/src/com/brashmonkey/spriter/Animation.java
@@ -87,7 +87,7 @@ public class Animation {
      * @param time The time which has to be between 0 and {@link #length} to work properly.
      * @param root The root bone which is not allowed to be null. The whole animation runs relative to the root bone.
      */
-    public void update(int time, Bone root){
+    public void update(float time, Bone root){
     	if(!this.prepared) throw new SpriterException("This animation is not ready yet to animate itself. Please call prepare()!");
     	if(root == null) throw new SpriterException("The root can not be null! Set a root bone to apply this animation relative to the root bone.");
     	this.currentKey = mainline.getKeyBeforeTime(time);
@@ -100,7 +100,7 @@ public class Animation {
 			this.update(ref, root, time);
     }
     
-    protected void update(BoneRef ref, Bone root, int time){
+    protected void update(BoneRef ref, Bone root, float time){
     	boolean isObject = ref instanceof ObjectRef;
 		//Get the timelines, the refs pointing to
 		Timeline timeline = getTimeline(ref.timeline);
